+++ scout/server/blueprints/variants/forms.py
@@ -187,12 +187,6 @@ class CancerSvFiltersForm(SvFiltersForm):
 
     depth = IntegerField("Depth >", validators=[validators.Optional()])
     alt_count = IntegerField("Min alt count", validators=[validators.Optional()])
-    control_frequency = BetterDecimalField(
-        "Normal alt AF <", places=2, validators=[validators.Optional()]
-    )
-    tumor_frequency = BetterDecimalField(
-        "Tumor alt AF >", places=2, validators=[validators.Optional()]
-    )
 
 
 FILTERSFORMCLASS = {